requirements.txt
Pillow==9.2.0
streamlit==1.12.2

